feat(card): moving adapter.js outside of the specs folder - TWIG-132
