+++ src/ec/packages/ec-component-card/.npmignore
@@ -1,5 +1,4 @@
 __snapshots__
 docs
-*.story.js
-*.test.js
+*.js
 **/*.md
