+++ src/ec/packages/ec-component-card/adapter.js
