+++ src/ec/packages/ec-component-card/demo/data--card-event.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/no-extraneous-dependencies, unicorn/filename-case */
 import specCardEvent from '@ecl/ec-specs-card/demo/data--card-event';
-import adapter from './adapter';
+import adapter from '../adapter';
 
 export default adapter(specCardEvent);
