+++ src/ec/packages/ec-component-card/demo/data--card-tag.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/no-extraneous-dependencies, unicorn/filename-case */
 import specCardTag from '@ecl/ec-specs-card/demo/data--card-tag';
-import adapter from './adapter';
+import adapter from '../adapter';
 
 export default adapter(specCardTag);
