+++ src/ec/packages/ec-component-card/demo/data--card.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/no-extraneous-dependencies, unicorn/filename-case */
 import specCard from '@ecl/ec-specs-card/demo/data--card';
-import adapter from './adapter';
+import adapter from '../adapter';
 
 export default adapter(specCard);
