+++ src/ec/packages/ec-component-card/demo/data--tile.js
@@ -1,6 +1,6 @@
 /* eslint-disable import/no-extraneous-dependencies, unicorn/filename-case */
 import specTile from '@ecl/ec-specs-card/demo/data--tile';
-import adapter from './adapter';
+import adapter from '../adapter';
 
 specTile.type = 'tile';
